# static/openapi.yaml
openapi: 3.1.0
info:
  title: My Manga List API - OpenAPI 3.1
  description: |-
    My Manga List API is a self-hosted platform for tracking manga and manga-like series you have read, are reading, or plan to read. It is designed for personal use but can also be deployed by anyone who wishes to manage their own reading list.
    
    As a self-hosted FOSS solution, you have full control over your data and privacy. Please consider whether hosting and managing your own database fits your needs. If you wish to contribute to this project or need help, the documentation and community are here for you.
    - [Project Github Page](https://github.com/ma-tecik/MyMangaList)
    ***
    This project relies on public APIs from MangaUpdates, Mangadex, and (maybe) MyAnimeList to provide up-to-date manga information. We sincerely thank these communities and recommend users to contribute to their databases to keep the ecosystem rich and accurate. By using My Manga List API, you acknowledge and agree to the Terms of Service of these third-party platforms.
    - [MangaUpdates](https://www.mangaupdates.com/), [API usage Policy](https://api.mangaupdates.com/#section/Acceptable-Use-Policy)
    - [Mangadex](https://mangadex.org/), [API usage Policy](https://api.mangadex.org/docs/#acceptable-usage-policy)
    - And if [MyAnimeList](https://myanimelist.net/) Client ID provided: [API License and Developer Agreement](https://myanimelist.net/static/apiagreement.html)
...
  contact:
...
  license:
...
servers:
...
paths:
...
  /series:
    get:
      operationId: getSeriesList
      summary: Retrieve series list
      description: Returns a list of series with optional filters for status, type, and tags.
      tags: [ Series ]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: perpage
          in: query
          description: for pagination 30-100
          schema:
            type: integer
            default: 30
            minimum: 30
            maximum: 100
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/Status'
        - name: type
          in: query
          description: Filters by exact series type; use 'all' to include all types.
          schema:
            type: string
            enum:
              - all
              - Manga
              - Manhwa
              - Manhua
              - OEL
              - Vietnamese
              - Malaysian
              - Indonesian
              - Novel
              - Artbook
              - Other
              - minor
            default: all
            description: "minor: includes everything except Manga, Manhwa and Manhua"
        - name: included
          in: query
          style: form
          explode: false
          description: |-
            Genres to include in results.
            For Josei, Seinen, Shoujo, Shounen; works matching any (OR logic) will be included, for other genres, works must match all (AND logic).
            Example: "Josei,Seinen,Action,Adventure" returns series that are Josei OR Seinen AND also have Action AND Adventure.
          schema:
            $ref: '#/components/schemas/GenresSearch'
        - name: excluded
          in: query
          style: form
          explode: false
          description: |-
            Genres to exclude from results (works containing any of these will be omitted).
            Example: "Hentai,Yaoi" will exclude any series with those genres.
          schema:
            $ref: '#/components/schemas/GenresSearch'
        - name: sort_by
          in: query
          schema:
            type: string
            enum:
              - rating-mu
              - rating-dex
              - rating-mal
              - title
              - time
            default: rating-mu
      responses:
        '200':
          description: Series list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseSeriesGet'
        '400':
          description: Bad request (invalid query)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseKO'
        '500':
          description: Internal server error, check logs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseKO'
    post:
      operationId: createSeries
      summary: Add a series entry
      description: Adds a new series to your manga list.
      tags: [ Series ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeriesDataEA'
      responses:
        '201':
          description: Series created
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    default: OK
                  data:
                    $ref: '#/components/schemas/Series'
                    required:
                      - id
                      - thumbnail_ext
                      - ids
                      - title
                      - type
                      - is_md
                      - status
                      - authors
        '400':
          description: Bad request (invalid query)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseKO'
        '409':
          description: Series already exists, use merge or update endpoint.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseMerge'
        '500':
          description: Internal server error, check logs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseKO'
        '502':
          description: Error while downloading series thumbnail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseKO'
...
  /series/{id}/ratings:
    patch:
      operationId: updateSeriesRatings
      summary: Update series ratings and vote counts
      description: Updates/overwrites ratings and vote counts of given series
      tags: [ Series ]
      parameters:
        - name: id
          in: path
          required: true
          description: Internal series ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_rating:
                  type: number
                  format: float
                  minimum: 1
                  maximum: 10
                mu_rating:
                  type: number
                  format: float
                  minimum: 1
                  maximum: 10
                mu_votes:
                  type: integer
                  minimum: 1
                dex_rating:
                  type: number
                  format: float
                  minimum: 1
                  maximum: 10
                dex_votes:
                  type: integer
                  minimum: 1
                mal_rating:
                  type: number
                  format: float
                  minimum: 1
                  maximum: 10
                mal_votes:
                  type: integer
                  minimum: 1
      responses:
        '204':
          description: Updated
        '400':
          description: Bad request (invalid query)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseKO'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseKO'
        '500':
          description: Internal server error, check logs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseKO'
...
  /external/series/id:
    get:
      operationId: externalSeriesID
      summary: Get the series_id of a series by external IDs
      description: Returns the series_id of a series based on given one or more MU, Dex, MAL, Bato, line IDs.
      tags: [ Series ]
      parameters:
        - name: mu
          in: query
          schema:
            type: string
        - name: dex
          in: query
          schema:
            type: string
        - name: mal
          in: query
          schema:
            type: integer
        - name: bato
          in: query
          schema:
            type: integer
        - name: line
          in: query
          schema:
            type: string
      responses:
        '200':
          description: series_id
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    default: OK
                  data:
                    type: integer
        '400':
          description: Bad request (invalid query)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseKO'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseKO'
        '409':
          description: Series conflict; manual merge required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseMerge'
        '500':
          description: Internal server error, check logs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseKO'
...
components:
  schemas:
...
    Status:
      type: string
      description: Reading status
      enum:
        - plan-to
        - reading
        - completed
        - one-shots
        - dropped
        - on-hold
        - ongoing
...
    GenresSearch:
      allOf:
        - type: array
          items:
            type: string
            enum:
              - nsfw
        - $ref: '#/components/schemas/Genres'
...
    responseKO:
      type: object
      properties:
        result:
          type: string
          default: KO
        error:
          type: string
          description: 'error message'
      required:
        - result
        - error
...
    responseSeriesGet:
      type: object
      properties:
        result:
          type: string
          default: OK
        page:
          type: integer
          default: 1
        data:
          type: array
          items:
            type: object
            allOf:
              - type: object
                properties:
                  id:
                    type: integer
                    description: internal ID of the series
                  thumbnail_ext:
                    type: string
                    description: extension of thumbnail (e.g., jpg, png)
              - $ref: '#/components/schemas/SeriesData'
              - type: object
                properties:
                  rating:
                    type: number
                    format: float
                    description: Average rating of the series (based on selected or joined source)
                  user_rating:
                    type: number
                    format: float
...
  securitySchemes:
    cookieAuth:
      in: cookie
      name: session
      type: apiKey
security:
  - cookieAuth: [ ]